code refact
