+++ src/Moonglade.Web/Program.cs
@@ -31,7 +31,7 @@ string connStr = builder.Configuration.GetConnectionString("MoongladeDatabase");
 
 var cultures = new[] { "en-US", "zh-Hans" }.Select(p => new CultureInfo(p)).ToList();
 
-WriteParameterTable();
+builder.WriteParameterTable();
 AnsiConsole.MarkupLine("[link=https://github.com/EdiWang/Moonglade]GitHub: EdiWang/Moonglade[/]");
 
 ConfigureConfiguration();
@@ -45,42 +45,6 @@ await FirstRun();
 ConfigureMiddleware();
 
 app.Run();
-
-void WriteParameterTable()
-{
-    var appVersion = Helper.AppVersion;
-    var table = new Table
-    {
-        Title = new($"Moonglade.Web {appVersion} | .NET {Environment.Version}")
-    };
-
-    var strHostName = Dns.GetHostName();
-    var ipEntry = Dns.GetHostEntry(strHostName);
-    var ips = ipEntry.AddressList;
-
-    var envName = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
-
-    table.AddColumn("Parameter");
-    table.AddColumn("Value");
-    table.AddRow(new Markup("[blue]Path[/]"), new Text(Environment.CurrentDirectory));
-    table.AddRow(new Markup("[blue]System[/]"), new Text(Helper.TryGetFullOSVersion()));
-    table.AddRow(new Markup("[blue]User[/]"), new Text(Environment.UserName));
-    table.AddRow(new Markup("[blue]Host[/]"), new Text(Environment.MachineName));
-    table.AddRow(new Markup("[blue]IP addresses[/]"), new Rows(ips.Select(p => new Text(p.ToString()))));
-    table.AddRow(new Markup("[blue]Database type[/]"), new Text(dbType!));
-
-    if (!string.IsNullOrWhiteSpace(envName) && envName.ToLower() == "development")
-    {
-        table.AddRow(new Markup("[blue]Connection String[/]"), new Text(builder.Configuration.GetConnectionString("MoongladeDatabase")));
-    }
-
-    table.AddRow(new Markup("[blue]Image storage[/]"), new Text(builder.Configuration["ImageStorage:Provider"]!));
-    table.AddRow(new Markup("[blue]Authentication provider[/]"), new Text(builder.Configuration["Authentication:Provider"]!));
-    table.AddRow(new Markup("[blue]Editor[/]"), new Text(builder.Configuration["Editor"]!));
-    table.AddRow(new Markup("[blue]ASPNETCORE_ENVIRONMENT[/]"), new Text(envName ?? "N/A"));
-
-    AnsiConsole.Write(table);
-}
 
 void ConfigureConfiguration()
 {
